Created different files for enums and helper methods
As mentioned in the PR review separate files for enums and it's extension functions are created. Also separate file to define response object is created. After testing once again, MaterialTapTargetSize property is removed as it was not required.
