+++ lib/assets/enums.dart
@@ -1,0 +1,11 @@
+enum Gender {
+  MALE,
+  FEMALE,
+  OTHERS,
+}
+
+enum Country {
+  INDIA,
+  CHINA,
+  ENGLAND,
+}
