+++ lib/assets/response_object.dart
@@ -1,0 +1,14 @@
+class Response {
+  String name;
+  String dob;
+  String? gender;
+  String opinion;
+  String? country;
+
+  Response(
+      {required this.name,
+      required this.dob,
+      required this.gender,
+      required this.opinion,
+      required this.country});
+}
